+++ dialect/psql/select_test.go
@@ -252,6 +252,39 @@ WINDOW w AS (PARTITION BY depname ORDER BY salary)`,
 			),
 			ExpectedSQL: `SELECT id, name FROM users UNION select id, name FROM admins UNION select id, name FROM mods`,
 		},
+		"Union with combined args": {
+			Query: psql.Select(
+				sm.Columns("id", "name"),
+				sm.From("users"),
+				sm.Limit(100),
+				sm.OrderBy("id"),
+				sm.Union(psql.Select(
+					sm.Columns("id", "name"),
+					sm.From("admins"),
+					sm.Limit(10),
+					sm.OrderBy("id"),
+				)),
+				sm.OrderCombined("id"),
+				sm.LimitCombined(1000),
+			),
+			ExpectedSQL: `(SELECT id, name FROM users ORDER BY id LIMIT 100) UNION (SELECT id, name FROM admins ORDER BY id LIMIT 10)
+ORDER BY id LIMIT 1000`,
+		},
+		"Union with uncombined args": {
+			Query: psql.Select(
+				sm.Columns("id", "name"),
+				sm.From("users"),
+				sm.Limit(1),
+				sm.OrderBy("id"),
+				sm.Union(psql.Select(
+					sm.Columns("id", "name"),
+					sm.From("admins"),
+					sm.Limit(1),
+					sm.OrderBy("id"),
+				)),
+			),
+			ExpectedSQL: `(SELECT id, name FROM users ORDER BY id LIMIT 1) UNION (SELECT id, name FROM admins ORDER BY id LIMIT 1)`,
+		},
 	}
 
 	testutils.RunTests(t, examples, formatter)